App completed with logic, modal left.
