+++ src/App.jsx
@@ -1,4 +1,4 @@
-import React, { useState } from "react";
+import React, { useEffect, useState } from "react";
 import Game from "./components/game";
 import Rules from "./components/rules";
 import Decider from "./components/decider";
@@ -10,15 +10,14 @@ function App() {
   const [houseChoice, setHouseChoice] = useState("");
   //state to keep track if user has played or not
   const [chosen, setChosen] = useState(false);
-
-  const handleScore = (updatedScore) => {
-    setScore(updatedScore);
-  };
+  const [winner, setWinner] = useState("");
+  const [score, setScore] = useState(0);
 
   //lifted state function to handle play again
   const handlePlayAgain = () => {
     setChosen(false);
     setUserChoice("");
+    setHouseChoice("");
   };
 
   //Function to handle user choice and update the states
@@ -29,6 +28,43 @@ function App() {
     const hands = ["rock", "paper", "scissor"];
     setHouseChoice(hands[Math.floor(Math.random() * hands.length)]);
   };
+
+  /*
+scenarios
+user    house    result
+paper   rock     user
+paper   scissor  house
+scissor rock     house
+scissor paper    user
+rock    paper    house
+rock    scissor  user
+*/
+  /*
+Function to decide winner 
+@params user's choice, house's choice
+returns outcome of the game
+*/
+  const declareWinner = (user, house) => {
+    if (user === house) {
+      setWinner("Draw");
+    } else if (
+      (user === "paper" && house === "rock") ||
+      (user === "scissor" && house === "paper") ||
+      (user === "rock" && house === "scissor")
+    ) {
+      setWinner("You Win");
+      setScore((prevScore) => (prevScore += 1));
+    } else {
+      setWinner("You Lose");
+      setScore((prevScore) =>
+        prevScore === 0 ? (prevScore = 0) : (prevScore -= 1)
+      );
+    }
+  };
+
+  useEffect(() => {
+    declareWinner(userChoice, houseChoice);
+  }, [userChoice]);
 
   return (
     <div className="container mx-auto">
@@ -45,7 +81,7 @@ function App() {
           <p className="absolute top-2 left-10 text-lg font-semibold uppercase tracking-wider">
             Score
           </p>
-          <h1 className="mt-2 text-[4rem] font-bold text-darkText">{2}</h1>
+          <h1 className="mt-2 text-[4rem] font-bold text-darkText">{score}</h1>
         </div>
       </div>
       <div className="h-[20rem]">
@@ -54,6 +90,7 @@ function App() {
             userChoice={userChoice}
             playAgain={handlePlayAgain}
             houseChoice={houseChoice}
+            winner={winner}
           />
         ) : (
           <Game handleUserChoice={handleUserChoice} />
